format

+++ fincore.c
@@ -116,7 +116,7 @@ void fincore(char* path,
 
         printf( "%-120s %15s\n",
                 path,
-                _ltoa( file_stat.st_size )
+                _ltoa( (long)file_stat.st_size )
                 
                 
                 );
